WIP. initial block-based board state storage implementation, server-side only

+++ wordland-server/src/main/java/wordland/model/game/GameStateChange.java
@@ -12,26 +12,26 @@ import wordland.model.support.PlayedTile;
 @NoArgsConstructor @Accessors(chain=true)
 public class GameStateChange {
 
-    public GameStateChange (int version, GameStateChangeType stateChange, Object object) {
+    public GameStateChange (long version, GameStateChangeType stateChange, Object object) {
         this.version = version;
         this.stateChange = stateChange;
         this.object = JsonUtil.FULL_MAPPER.valueToTree(object);
     }
 
-    @Getter @Setter private int version;
+    @Getter @Setter private long version;
     @Getter @Setter private String room;
     @Getter @Setter private GameStateChangeType stateChange;
     @Getter @Setter private JsonNode object;
 
-    public static GameStateChange playerJoined(int version, GamePlayer player) {
+    public static GameStateChange playerJoined(long version, GamePlayer player) {
         return new GameStateChange(version, GameStateChangeType.player_joined, player);
     }
 
-    public static GameStateChange playerLeft(int version, String id) {
+    public static GameStateChange playerLeft(long version, String id) {
         return new GameStateChange(version, GameStateChangeType.player_left, id);
     }
 
-    public static GameStateChange wordPlayed(int version, GamePlayer player, PlayedTile[] tiles) {
+    public static GameStateChange wordPlayed(long version, GamePlayer player, PlayedTile[] tiles) {
         return new GameStateChange(version, GameStateChangeType.word_played, new GameRuntimeEvent().setId(player.getId()).setTiles(tiles));
     }
